requirements.txt
amqp==5.1.1
asgiref==3.5.2
billiard==3.6.4.0
braintree==4.16.0
Brotli==1.0.9
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cssselect2==0.6.0
defusedxml==0.7.1
Django==3.2.10
django-allauth==0.41.0
django-cors-headers==3.1.1
django-rest-auth==0.9.5
djangorestframework==3.11.0
fonttools==4.35.0
html5lib==1.1
idna==3.3
kombu==5.2.4
oauthlib==3.2.0
Pillow==7.1.0
prompt-toolkit==3.0.30
psycopg2-binary==2.8.5
pycparser==2.21
pydyf==0.2.0
pyphen==0.12.0
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
tinycss2==1.1.1
urllib3==1.26.11
vine==5.0.0
wcwidth==0.2.5
weasyprint==56.1
webencodings==0.5.1
zopfli==0.2.1
gunicorn==20.1.0
whitenoise==6.2.0
dj-database-url==1.0.0
python-decouple==3.6